Change the API of INIFile - it now throws UncheckedIOException if it can't load or save the file, and IllegalStateException if the file is wrongly-formatted.

+++ src/test/java/org/devzendo/commoncode/file/TestINIFile.java
@@ -16,17 +16,16 @@
 
 package org.devzendo.commoncode.file;
 
-import java.io.File;
-import java.io.IOException;
-
+import org.apache.log4j.Logger;
 import org.devzendo.commoncode.logging.LoggingUnittestHelper;
-import org.junit.After;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
+import org.hamcrest.text.MatchesPattern;
+import org.junit.*;
+import org.junit.rules.ExpectedException;
 
-
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.io.UncheckedIOException;
 
 /**
  * Tests the INIFile (not written TDD)
@@ -34,13 +33,15 @@ import org.junit.Test;
  *
  */
 public final class TestINIFile {
+    private static final Logger LOGGER = Logger.getLogger(TestINIFile.class);
+
     private static final String SECTION_ARRAY = "ARRAY";
     private INIFile iniFile;
     private File tempFile;
 
-    /**
-     * 
-     */
+    @Rule
+    public ExpectedException thrown = ExpectedException.none();
+
     @BeforeClass
     public static void setupLogging() {
         LoggingUnittestHelper.setupLogging();
@@ -50,7 +51,6 @@ public final class TestINIFile {
      * Create temporary ini file
      * @throws IOException on failure
      */
-    @Before
     public void getINIFile() throws IOException {
         tempFile = File.createTempFile("common-unit-test", "ini").getAbsoluteFile();
         tempFile.deleteOnExit();
@@ -63,14 +63,18 @@ public final class TestINIFile {
      */
     @After
     public void deleteTempFile() throws IOException {
-        tempFile.delete();
+        if (tempFile != null) {
+            tempFile.delete();
+        }
     }
 
     /**
      * It's OK if your array is empty
      */
     @Test
-    public void setEmptyArrayOK() {
+    public void setEmptyArrayOK() throws IOException {
+        getINIFile();
+
         iniFile.setArray(SECTION_ARRAY, new String[0]);
         final String[] array = iniFile.getArray(SECTION_ARRAY);
         Assert.assertNotNull(array);
@@ -81,7 +85,9 @@ public final class TestINIFile {
      * An array should be stored and retrieved in the same order
      */
     @Test
-    public void arraysShouldBeStoredAndRetrievedOrdered() {
+    public void arraysShouldBeStoredAndRetrievedOrdered() throws IOException {
+        getINIFile();
+
         final String[] one = new String[] {
                 "one", "two", "three", "four"
         };
@@ -99,7 +105,9 @@ public final class TestINIFile {
      * truncates the array.
      */
     @Test
-    public void testSetArrayTruncates() {
+    public void testSetArrayTruncates() throws IOException {
+        getINIFile();
+
         final String[] one = new String[] {
                 "one", "two", "three", "four"
         };
@@ -116,27 +124,90 @@ public final class TestINIFile {
         }
     }
     
-    /**
-     * 
-     */
     @Test(expected = IllegalArgumentException.class)
-    public void nullSectionIsDisallowed() {
+    public void nullSectionIsDisallowed() throws IOException {
+        getINIFile();
+
         iniFile.setValue(null, "foo", "bar");
     }
 
-    /**
-     * 
-     */
     @Test(expected = IllegalArgumentException.class)
-    public void nullNameIsDisallowed() {
+    public void nullNameIsDisallowed() throws IOException {
+        getINIFile();
+
         iniFile.setValue("section", null, "bar");
     }
 
-    /**
-     * 
-     */
     @Test(expected = IllegalArgumentException.class)
-    public void nullValueIsDisallowed() {
+    public void nullValueIsDisallowed() throws IOException {
+        getINIFile();
+
         iniFile.setValue("section", "name", null);
+    }
+
+    /**
+     * Prior to 1.1.0, failure to save and load was logged but then silently swallowed. In 1.1.0, the
+     * IOExceptions are converted to UncheckedIOExceptions.
+     */
+    @Test
+    public void loadFailureDueToFileNotFound() throws IOException {
+        thrown.expect(UncheckedIOException.class);
+        thrown.expectMessage(MatchesPattern.matchesPattern("INI file .* not found"));
+
+        // try to load from a 'file' that's actually a directory
+        File itsADirectory = new File(getTempDir(), "its-a-directory").getAbsoluteFile();
+        LOGGER.info("directory is " + itsADirectory.getAbsolutePath());
+        itsADirectory.deleteOnExit();
+        Assert.assertTrue(itsADirectory.mkdir());
+        Assert.assertTrue(itsADirectory.isDirectory());
+        Assert.assertTrue(itsADirectory.exists());
+
+        new INIFile(itsADirectory.getAbsolutePath());
+    }
+
+    @Test
+    public void nameValuePairNotUnderSectionThrows() throws IOException {
+        thrown.expect(IllegalStateException.class);
+        thrown.expectMessage(MatchesPattern.matchesPattern("Line 1 name=value line not under any \\[section\\]: 'foo=bar'"));
+
+        tempFile = writeToTempFile("foo=bar");
+
+        new INIFile(tempFile.getAbsolutePath());
+    }
+
+    @Test
+    public void lineNotNameValuePairNorSectionThrows() throws IOException {
+        thrown.expect(IllegalStateException.class);
+        thrown.expectMessage(MatchesPattern.matchesPattern("Line 1 not matched against \\[section\\] or name=value: 'barafundle'"));
+
+        tempFile = writeToTempFile("barafundle");
+
+        new INIFile(tempFile.getAbsolutePath());
+    }
+
+    @Test
+    public void saveFailureDueToIOException() throws IOException {
+        getINIFile();
+        Assert.assertTrue(tempFile.setWritable(false));
+
+        thrown.expect(UncheckedIOException.class);
+        thrown.expectMessage(MatchesPattern.matchesPattern("Could not write INI file .*: .*"));
+
+        iniFile.setValue("foo", "bar", "quux");
+    }
+
+    private File writeToTempFile(final String... strings) throws IOException {
+        final File t =  File.createTempFile("common-unit-test", "ini").getAbsoluteFile();
+        t.deleteOnExit();
+        try (final FileWriter ft = new FileWriter(t)) {
+            for (String string: strings) {
+                ft.write(string + System.lineSeparator());
+            }
+        };
+        return t;
+    }
+
+    private File getTempDir() {
+        return new File(System.getProperty("java.io.tmpdir"));
     }
 }
